+++ resources/lang/en/admin/moon.php
@@ -1,6 +1,15 @@
 <?php
 
 return [
+    'create' => [
+        'mother' => 'ID of mother planet',
+        'name' => 'Name of the moon',
+        'submit' => 'Create',
+        'success_text' => 'Moon has been created successfully',
+        'success_title' => 'Created',
+        'table_title' => 'Creation form',
+        'title' => 'Give a moon',
+    ],
     'index' => [
         'id' => 'ID',
         'galaxy' => 'Galaxy',