+++ resources/lang/en/app.php
@@ -1,6 +1,7 @@
 <?php
 
 return [
+    'default_moon_name' => 'Moon',
     'error' => 'Error',
     'format-datetime' => 'd/m/Y h:i:s',
     'name' => 'Blackout',